fix: move sg node out
